+++ crates/matrix-sdk/src/paginators/room.rs
@@ -49,11 +49,11 @@ pub enum PaginatorState {
 
 /// Paginations tokens used for backward and forward pagination.
 #[derive(Debug)]
-struct PaginationTokens {
+pub(crate) struct PaginationTokens {
     /// Pagination token used for backward pagination.
-    previous: PaginationToken,
+    pub(crate) previous: PaginationToken,
     /// Pagination token used for forward pagination.
-    next: PaginationToken,
+    pub(crate) next: PaginationToken,
 }
 
 /// A stateful object to reach to an event, and then paginate backward and
@@ -338,6 +338,16 @@ impl<PR: PaginableRoom> Paginator<PR> {
         self.state.set(PaginatorState::Idle);
 
         Ok(PaginationResult { events: response.chunk, hit_end_of_timeline })
+    }
+
+    /// Returns the current `prev_token` used for pagination.
+    pub fn prev_token(&self) -> PaginationToken {
+        self.tokens.lock().unwrap().previous.clone()
+    }
+
+    /// Returns the current `next_token` used for pagination.
+    pub fn next_token(&self) -> PaginationToken {
+        self.tokens.lock().unwrap().next.clone()
     }
 }